Update symlinks scripts.
init-symlinks made an error when a path already existed but was not a symlink:
it added the symlink to the existing directory. Now it will deleted the existing
path, so be carefull (use version control).

Also, check-symlinks is a new script to check:

- wether paths in a submodule do not exist in the current module yet, and
- wether lines in the .symlinks file are actually existing paths.

Both should prevent cruft building up and missing paths.